requirements.txt
pandas
numpy
pyarrow
scipy
pytest
matplotlib
plotly
pyyaml